Add tegere.query/query function
It queries a features data structure (conformant to ``:tegere.parser/features``)
using a clause ``where`` and returning a substructure of ``features`` that
contains only the features and scenarios that match the where clause. The
``where`` clause is a list-based tree structure that conforms to
``:tegere.query/query-tree``, i.e., its non-terminal nodes are the symbols 'and
'or or 'not and its terminal nodes are strings that should match Gherkin tags
in the features.